Fix memory leak
There was a memory leak due to a bug in the SymbolicExecutor
class. The cause of the leak was a circular dependency between
the SymbolicExecutor and the CallbackManager classes. The latter
binds to the former and never unbinds. The solution was to add
the unbind method and to call at the end of the run method in
the SymbolicExecutor class.